1312
Added small red "X" buttons to the player item display for admins to remove items.
Updated the health display to use a visual bar that changes color based on health.
Added functionality to display item details in a modal when clicking on item images.
Adjusted item display to include the health bar.